new Server error fixed

+++ server_setup/KqueueIdent.hpp
@@ -20,15 +20,15 @@ class KqueueIdent
 {
 	public :
 
-		virtual void			set_ident( const uintptr_t & id ) = 0;
+		virtual void			set_ident( const int & id ) = 0;
 		virtual void			set_type( const t_ident_type & type ) = 0;
 
-		virtual uintptr_t		get_ident( void ) = 0;
+		virtual int				get_ident( void ) = 0;
 		virtual t_ident_type	get_type( void ) = 0;
 
 	protected :
 
-		uintptr_t		_ident;
+		int				_ident;
 		t_ident_type	_type;
 };
